# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Documents API
  description: |
    API for managing and processing web documents with background processing support.
    
    ## Features
    - Idempotent document creation
    - Background processing with status tracking
    - Document lifecycle management
    - Vector store integration for content retrieval
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /documents:
    post:
      tags:
        - Documents
      summary: Create a new document
      description: |
        Create a new document for processing. This endpoint is idempotent - 
        calling it multiple times with the same user_id and URL will return 
        the existing document instead of creating duplicates.
        
        The document is automatically queued for background processing.
      operationId: createDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentCreate'
            examples:
              basic:
                summary: Basic document creation
                value:
                  user_id: user123
                  url: https://example.com/article
              full:
                summary: Document with all fields
                value:
                  user_id: user123
                  url: https://example.com/article
                  title: Example Article
                  description: An interesting article about technology
                  dom: "<html>...</html>"
      responses:
        '201':
          description: Document created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
              example:
                _id: 507f1f77bcf86cd799439011
                user_id: user123
                url: https://example.com/article
                dom: null
                title: Example Article
                description: An interesting article
                summary: null
                processing_status: QUEUED
                created_at: "2025-10-18T10:30:00Z"
                updated_at: "2025-10-18T10:30:00Z"
        '200':
          description: Document already exists (idempotent response)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Documents
      summary: Get all documents
      description: |
        Retrieve all documents for a specific user with pagination support.
        Documents are returned in reverse chronological order (newest first).
      operationId: getAllDocuments
      parameters:
        - name: user_id
          in: query
          required: true
          description: User ID to filter documents
          schema:
            type: string
          example: user123
        - name: skip
          in: query
          required: false
          description: Number of documents to skip (for pagination)
          schema:
            type: integer
            minimum: 0
            default: 0
          example: 0
        - name: limit
          in: query
          required: false
          description: Maximum number of documents to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          example: 20
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
              example:
                - _id: 507f1f77bcf86cd799439011
                  user_id: user123
                  url: https://example.com/article1
                  dom: null
                  title: First Article
                  description: Description of first article
                  summary: This article provides an overview of key concepts in software development, focusing on best practices and modern approaches to building scalable applications.
                  processing_status: COMPLETE
                  created_at: "2025-10-18T10:30:00Z"
                  updated_at: "2025-10-18T10:35:00Z"
                - _id: 507f1f77bcf86cd799439012
                  user_id: user123
                  url: https://example.com/article2
                  dom: null
                  title: Second Article
                  description: Description of second article
                  summary: null
                  processing_status: IN_PROGRESS
                  created_at: "2025-10-18T11:00:00Z"
                  updated_at: "2025-10-18T11:05:00Z"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /documents/{document_id}:
    get:
      tags:
        - Documents
      summary: Get document by ID
      description: Retrieve a specific document by its ID
      operationId: getDocumentById
      parameters:
        - name: document_id
          in: path
          required: true
          description: MongoDB ObjectId of the document
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Document found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
              example:
                _id: 507f1f77bcf86cd799439011
                user_id: user123
                url: https://example.com/article
                dom: null
                title: Example Article
                description: An example article description
                summary: This article explores the fundamentals of machine learning and its applications in modern technology, including practical examples and use cases.
                processing_status: COMPLETE
                created_at: "2025-10-18T10:30:00Z"
                updated_at: "2025-10-18T10:35:00Z"
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: Document not found
        '400':
          description: Invalid document ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Documents
      summary: Delete document
      description: Delete a specific document by its ID
      operationId: deleteDocument
      parameters:
        - name: document_id
          in: path
          required: true
          description: MongoDB ObjectId of the document
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Document deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Document deleted successfully
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: Document not found
        '400':
          description: Invalid document ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /summary:
    post:
      tags:
        - Processing
      summary: Get summary of web resource
      description: Generate and retrieve a summary of the specified web resource
      operationId: getSummary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebResource'
            example:
              user_id: user123
              access_token: token_abc123
              web_url: https://example.com/article
      responses:
        '200':
          description: Summary generated successfully
          content:
            application/json:
              schema:
                type: object
                description: Summary response (structure may vary)
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /index/webresource:
    post:
      tags:
        - Processing
      summary: Index web resource
      description: Index a web resource for vector store retrieval
      operationId: indexWebResource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebResource'
            example:
              user_id: user123
              access_token: token_abc123
              web_url: https://example.com/article
      responses:
        '200':
          description: Resource indexed successfully
          content:
            application/json:
              schema:
                type: object
                description: Indexing response (structure may vary)
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /retrieve/query:
    post:
      tags:
        - Processing
      summary: Query vector store
      description: Query the vector store for relevant content (returns top 2 results)
      operationId: queryVectorStore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebResource'
            example:
              user_id: user123
              access_token: token_abc123
              web_url: https://example.com/article
      responses:
        '200':
          description: Query executed successfully
          content:
            application/json:
              schema:
                type: object
                description: Query results (structure may vary)
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /categories:
    post:
      tags:
        - Categories
      summary: Create a new category
      description: Create a new category for organizing documents. Category names must be unique per user.
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
            examples:
              basic:
                summary: Basic category creation
                value:
                  user_id: user123
                  name: Technology
              full:
                summary: Category with description
                value:
                  user_id: user123
                  name: Technology
                  description: Articles about technology and software development
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
              example:
                _id: 507f1f77bcf86cd799439011
                user_id: user123
                name: Technology
                description: Articles about technology and software development
                document_ids: []
                created_at: "2025-10-18T10:30:00Z"
                updated_at: "2025-10-18T10:30:00Z"
        '409':
          description: Category with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "Category with name 'Technology' already exists for this user"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Categories
      summary: Get all categories
      description: Retrieve all categories for a specific user with pagination support.
      operationId: getAllCategories
      parameters:
        - name: user_id
          in: query
          required: true
          description: User ID to filter categories
          schema:
            type: string
          example: user123
        - name: skip
          in: query
          required: false
          description: Number of categories to skip (for pagination)
          schema:
            type: integer
            minimum: 0
            default: 0
          example: 0
        - name: limit
          in: query
          required: false
          description: Maximum number of categories to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          example: 20
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
              example:
                - _id: 507f1f77bcf86cd799439011
                  user_id: user123
                  name: Technology
                  description: Tech articles
                  document_ids: ["507f1f77bcf86cd799439015", "507f1f77bcf86cd799439016"]
                  created_at: "2025-10-18T10:30:00Z"
                  updated_at: "2025-10-18T10:35:00Z"
                - _id: 507f1f77bcf86cd799439012
                  user_id: user123
                  name: Business
                  description: Business news
                  document_ids: []
                  created_at: "2025-10-18T11:00:00Z"
                  updated_at: "2025-10-18T11:00:00Z"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /categories/{category_id}:
    get:
      tags:
        - Categories
      summary: Get category by ID
      description: Retrieve a specific category by its ID
      operationId: getCategoryById
      parameters:
        - name: category_id
          in: path
          required: true
          description: MongoDB ObjectId of the category
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        - name: user_id
          in: query
          required: true
          description: User ID for authorization
          schema:
            type: string
          example: user123
      responses:
        '200':
          description: Category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
              example:
                _id: 507f1f77bcf86cd799439011
                user_id: user123
                name: Technology
                description: Tech articles
                document_ids: ["507f1f77bcf86cd799439015"]
                created_at: "2025-10-18T10:30:00Z"
                updated_at: "2025-10-18T10:35:00Z"
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: Category not found
        '403':
          description: Not authorized to access this category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: Not authorized to access this category
        '400':
          description: Invalid category ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Categories
      summary: Update category
      description: Update category name or description (rename category)
      operationId: updateCategory
      parameters:
        - name: category_id
          in: path
          required: true
          description: MongoDB ObjectId of the category
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        - name: user_id
          in: query
          required: true
          description: User ID for authorization
          schema:
            type: string
          example: user123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
            examples:
              rename:
                summary: Rename category
                value:
                  name: Tech & Innovation
              description:
                summary: Update description
                value:
                  description: Updated description
              both:
                summary: Update both fields
                value:
                  name: Tech & Innovation
                  description: Articles about technology, innovation, and software
      responses:
        '200':
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Category with new name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Categories
      summary: Delete category
      description: Delete a category. Note that documents in the category are not deleted, only the category itself.
      operationId: deleteCategory
      parameters:
        - name: category_id
          in: path
          required: true
          description: MongoDB ObjectId of the category to delete
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        - name: user_id
          in: query
          required: true
          description: User ID for authorization
          schema:
            type: string
          example: user123
      responses:
        '200':
          description: Category deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Category deleted successfully
        '404':
          description: Category not found or user not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: Category not found
        '400':
          description: Invalid category ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Document:
      type: object
      description: Document object representing a web URL with processing status
      required:
        - user_id
        - url
        - processing_status
      properties:
        _id:
          type: string
          description: MongoDB ObjectId as string
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        user_id:
          type: string
          description: User ID who owns the document
          example: user123
        url:
          type: string
          format: uri
          description: Web URL of the document
          example: https://example.com/article
        dom:
          type: string
          nullable: true
          description: DOM content if available
          example: "<html>...</html>"
        title:
          type: string
          nullable: true
          description: Document title
          example: Example Article
        description:
          type: string
          nullable: true
          description: Document description
          example: An interesting article about technology
        summary:
          type: string
          nullable: true
          description: AI-generated summary of the document content (populated after processing)
          example: This article discusses the latest developments in artificial intelligence and machine learning, covering key trends and future implications.
        processing_status:
          $ref: '#/components/schemas/ProcessingStatus'
        created_at:
          type: string
          format: date-time
          description: Document creation timestamp (ISO 8601)
          example: "2025-10-18T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Document last update timestamp (ISO 8601)
          example: "2025-10-18T10:35:00Z"
    DocumentCreate:
      type: object
      description: Request body for creating a new document
      required:
        - user_id
        - url
      properties:
        user_id:
          type: string
          description: User ID who owns the document
          example: user123
        url:
          type: string
          format: uri
          description: Web URL of the document
          example: https://example.com/article
        dom:
          type: string
          nullable: true
          description: DOM content (optional)
          example: "<html>...</html>"
        title:
          type: string
          nullable: true
          description: Document title (optional)
          example: Example Article
        description:
          type: string
          nullable: true
          description: Document description (optional)
          example: An interesting article about technology
...
    WebResource:
      type: object
      description: Web resource for processing operations
      required:
        - web_url
      properties:
        user_id:
          type: string
          nullable: true
          description: User ID (optional)
          example: user123
        access_token:
          type: string
          nullable: true
          description: Access token for authentication (optional)
          example: token_abc123
        web_url:
          type: string
          format: uri
          description: Web URL to process
          example: https://example.com/article
...
    CategoryUpdate:
      type: object
      description: Request body for updating a category
      properties:
        name:
          type: string
          nullable: true
          description: New category name (optional)
          example: Tech & Innovation
        description:
          type: string
          nullable: true
          description: New category description (optional)
          example: Updated description
...
    Error:
      type: object
      description: Error response
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message describing what went wrong
          example: Document not found
...